refactor(ableset): [mechanical] split ack-store into state/ableset_ack.rs
Pure code motion, no behavior change: AbleSetMismatchAck/AckMap and the
load/save/acknowledge/unacknowledge methods move from ableset_mismatch.rs
to a new sibling ableset_ack.rs (same precedent as background_tasks.rs).
ableset_mismatch.rs keeps only the pure comparison logic
(compute_ableset_mismatches/mismatch_for_number).

This is prep for #655 F4's domain refusal enum and F9's mutex/prune/cap/
validation, which land on ableset_ack.rs in the next commit — splitting
first keeps that diff readable instead of interleaving moved and changed
lines (per the issue's posted implementation-plan comment).
